PathEditor関係のコードを複数のファイルに分けた
